refactor: move `ConversationToggles` to separate file
